refactor: N+1 쿼리 최적화 및 GooglePlaceService 클래스 분리
- PlaceService.getPlaceStats() N+1 쿼리 문제 해결
- GooglePlaceService 단일 책임 원칙 적용 클래스 분리
- 테스트 코드 배치 쿼리 방식으로 업데이트
